api2/requirements.txt
flask
flask-smorest
flask-cors
flask-jwt-extended
supabase
python-dotenv
firebase-admin
gunicorn
